@@ -12,29 +12,22 @@ import { ContactComponent } from './containers/contact/contact.component';
 import { NavToggleComponent } from './components/nav-toggle/nav-toggle.component';
 import { InfoComponent } from './components/info/info.component';
 import { InfoItemComponent } from './components/info-item/info-item.component';
-import {HttpClientModule} from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import {AppRoutingModule} from './app-routing.module.ts.module';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    NavToggleComponent,
-    AboutComponent,
-    SkillsComponent,
-    ScrollSpyDirective,
-    HomeComponent,
-    ExperienceComponent,
-    EducationComponent,
-    ContactComponent,
-    InfoComponent,
-    InfoItemComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        NavToggleComponent,
+        AboutComponent,
+        SkillsComponent,
+        ScrollSpyDirective,
+        HomeComponent,
+        ExperienceComponent,
+        EducationComponent,
+        ContactComponent,
+        InfoComponent,
+        InfoItemComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
